Add files via upload
In order for this to run, you need to set the correct permissions when connecting to mariadb(user, password) and you need to have the css in the working directory, as well as have the correct path with the html. I couldn't figure out clickable links and putting many plots in a html file, so I don't have a lot of the plots necessary. I'm also not sure why I get an error with the main function, I think it has to do with M1.